# api/swagger.yaml
openapi: 3.0.3
info:
  title: SSW Logs Capture API
  description: |
    Enterprise-grade log capture, processing and distribution system.

    ## Features
    - **Multi-source log collection** (files and containers)
    - **Real-time processing** with configurable pipelines
    - **High-performance** async batching and delivery
    - **Enterprise security** with RBAC and audit logging
    - **Advanced observability** with metrics, tracing and SLO monitoring
    - **Resilient architecture** with circuit breakers, DLQ and backpressure

    ## Authentication
    The API supports multiple authentication methods:
    - **Basic Authentication**: Username and password
    - **Bearer Token**: API tokens
    - **JWT**: JSON Web Tokens (when configured)

    ## Rate Limiting
    API endpoints are rate-limited to ensure system stability.
    Default limits: 1000 requests/minute per client.

  version: "2.0.0"
  contact:
    name: SSW Logs Capture Support
    url: https://github.com/ssw/log_capturer_go
    email: support@ssw.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8401
    description: Development server
  - url: https://logs.company.com
    description: Production server
security:
  - BasicAuth: []
  - BearerAuth: []
paths:
...
  /stats:
    get:
      tags:
        - Monitoring
      summary: Application statistics
      description: |
        Returns detailed statistics about the application's performance and state.

        **Statistics include:**
        - Dispatcher queue status and throughput
        - Monitor statistics (files/containers watched)
        - Sink delivery performance
        - Position management status
        - Resource utilization

      operationId: getStats
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Application statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /config:
    get:
      tags:
        - Configuration
      summary: Current configuration
      description: |
        Returns the current application configuration (sanitized).

        **Security Note:** Sensitive information like passwords and tokens are redacted.

      operationId: getConfig
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /config/reload:
    post:
      tags:
        - Configuration
      summary: Reload configuration
      description: |
        Triggers a configuration reload without restarting the application.

        **Requirements:**
        - Hot reload must be enabled in configuration
        - User must have admin permissions

        **Behavior:**
        - Validates new configuration before applying
        - Gracefully updates components that support hot reload
        - Reports which components require restart

      operationId: reloadConfig
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Configuration reloaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Configuration reloaded successfully"
                  changes:
                    type: array
                    items:
                      type: object
                      properties:
                        component:
                          type: string
                        action:
                          type: string
                          enum: [updated, restart_required, no_change]
                        details:
                          type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Hot reload not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /positions:
    get:
      tags:
        - Monitoring
      summary: Position tracking status
      description: |
        Returns current position tracking information for files and containers.

        **Use cases:**
        - Monitor log processing progress
        - Verify exactly-once delivery
        - Debug position tracking issues
        - Capacity planning

      operationId: getPositions
      security:
        - BasicAuth: []
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by position type
          schema:
            type: string
            enum: [file, container]
        - name: limit
          in: query
          description: Maximum number of positions to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: source_id
          in: query
          description: Filter by specific source ID
          schema:
            type: string
      responses:
        '200':
          description: Position tracking status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dlq/stats:
    get:
      tags:
        - Monitoring
      summary: Dead Letter Queue statistics
      description: |
        Returns statistics about the Dead Letter Queue (DLQ).

        **DLQ contains:**
        - Logs that failed processing after all retries
        - Logs with invalid timestamps or format
        - Logs rejected by sinks

        **Use cases:**
        - Monitor data quality issues
        - Identify configuration problems
        - Plan reprocessing activities

      operationId: getDLQStats
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: DLQ statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DLQStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /slo/status:
    get:
      tags:
        - SLO
      summary: SLO monitoring status
      description: |
        Returns current Service Level Objective (SLO) status and error budget information.

        **SLO Types:**
        - **Availability**: Service uptime percentage
        - **Latency**: Processing time percentiles (P95, P99)
        - **Throughput**: Logs processed per second
        - **Error Rate**: Percentage of failed operations

      operationId: getSLOStatus
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: SLO status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SLOStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: SLO monitoring not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /goroutines/stats:
    get:
      tags:
        - Monitoring
      summary: Goroutine tracking statistics
      description: |
        Returns goroutine leak detection and resource usage statistics.

        **Includes:**
        - Current goroutine count vs baseline
        - Suspected leak detection
        - Memory usage correlation
        - Resource cleanup status

      operationId: getGoroutineStats
      security:
        - BasicAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Goroutine statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoroutineStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Goroutine tracking not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /security/audit:
    get:
      tags:
        - Security
      summary: Security audit log
      description: |
        Returns recent security events and audit log entries.

        **Events include:**
        - Authentication attempts (success/failure)
        - Authorization decisions
        - Configuration changes
        - Security policy violations

      operationId: getSecurityAudit
      security:
        - BasicAuth: []
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of entries to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: since
          in: query
          description: Return entries since this timestamp
          schema:
            type: string
            format: date-time
        - name: event_type
          in: query
          description: Filter by event type
          schema:
            type: string
            enum: [auth, authz, config, violation]
        - name: severity
          in: query
          description: Filter by severity level
          schema:
            type: string
            enum: [info, warning, error, critical]
      responses:
        '200':
          description: Security audit log
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityAuditResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Security auditing not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Basic HTTP authentication using username and password
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Authentication required"
            code: "AUTH_REQUIRED"
            message: "Valid credentials must be provided"
            timestamp: "2024-01-15T10:30:00Z"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Access denied"
            code: "INSUFFICIENT_PERMISSIONS"
            message: "User does not have required permissions for this operation"
            timestamp: "2024-01-15T10:30:00Z"
  schemas:
...
    StatsResponse:
      type: object
      properties:
        application:
          type: object
          properties:
            name:
              type: string
            version:
              type: string
            uptime:
              type: string
        dispatcher:
          type: object
          properties:
            total_processed:
              type: integer
              format: int64
            error_count:
              type: integer
              format: int64
            queue_size:
              type: integer
            queue_utilization:
              type: number
              format: float
            duplicates_detected:
              type: integer
              format: int64
            throughput_per_sec:
              type: number
              format: float
        file_monitor:
          type: object
          properties:
            files_watched:
              type: integer
            total_logs:
              type: integer
              format: int64
            errors:
              type: integer
              format: int64
        container_monitor:
          type: object
          properties:
            containers_monitored:
              type: integer
            total_logs:
              type: integer
              format: int64
            reconnections:
              type: integer
        sinks:
          type: object
          additionalProperties:
            type: object
            properties:
              queue_util:
                type: number
                format: float
              errors:
                type: integer
              last_success:
                type: string
                format: date-time
    ConfigResponse:
      type: object
      properties:
        app:
          type: object
        metrics:
          type: object
        processing:
          type: object
        dispatcher:
          type: object
        sinks:
          type: object
        # Note: Sensitive configuration is redacted
    PositionsResponse:
      type: object
      properties:
        files:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              position:
                type: integer
                format: int64
              file_size:
                type: integer
                format: int64
              last_update:
                type: string
                format: date-time
              inode:
                type: integer
                format: int64
        containers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              last_timestamp:
                type: string
                format: date-time
              stream:
                type: string
                enum: [stdout, stderr]
        summary:
          type: object
          properties:
            total_files:
              type: integer
            total_containers:
              type: integer
            last_flush:
              type: string
              format: date-time
    DLQStatsResponse:
      type: object
      properties:
        total_entries:
          type: integer
        size_mb:
          type: number
          format: float
        oldest_entry:
          type: string
          format: date-time
        newest_entry:
          type: string
          format: date-time
        retry_queue_size:
          type: integer
        error_categories:
          type: object
          additionalProperties:
            type: integer
        hourly_growth:
          type: array
          items:
            type: object
            properties:
              hour:
                type: string
                format: date-time
              entries:
                type: integer
...
    GoroutineStatsResponse:
      type: object
      properties:
        current:
          type: integer
        baseline:
          type: integer
        max_seen:
          type: integer
        tracked:
          type: integer
        suspected_leaks:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [healthy, warning, critical]
        last_check:
          type: string
          format: date-time
        uptime:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Detailed error description
        details:
          type: object
          description: Additional error context
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
          description: Request trace ID for debugging
...